Сделала разбиение кода на модули
