+++ js/data.js
@@ -1,0 +1,34 @@
+import { randomNumber } from './util';
+
+const NAMES = [
+  'Морти',
+  'Шаман',
+  'Каспер',
+  'Миха',
+  'Вайс',
+  'Грей',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTIONS = [
+  'Отдыхаю с котиком.',
+  'Мой песель самый лучший',
+  'Морти - собака сутулая',
+  'Не нравится - не смотри!',
+  'Жизнь - штука тяжелая!',
+  'Покойтесь с миром обои...',
+];
+
+const maxCommentLength = 140;
+
+const getRandomArrayElement = (elements) => elements[randomNumber(0, elements.length - 1)];
+
+export {NAMES, MESSAGES, DESCRIPTIONS, getRandomArrayElement, maxCommentLength};
